# server/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: Bible API
  description: |
    Complete Bible API with Role-Based Access Control (RBAC), multi-version support, RSS feeds, user management, file locations, and backup system.
    
    ## Role-Based Access Control
    - **Admin**: Full system access with all permissions
    - **User**: Standard user with Bible reading and format access
    - **Viewer**: Read-only access to Bible content
    - **Editor**: Can edit content but not manage users
    
    ## Authentication
    Use Bearer token authentication. Login to get your JWT token.
    
    ## Rate Limiting
    - General API: 100 requests per 15 minutes
    - Auth endpoints: 5 attempts per 15 minutes
    
    ## File Locations
    Users can configure local drive paths where TXT and XML files will be saved automatically when accessing verses.
  version: 3.0.0
  contact:
    name: API Support
    email: support@bibleapi.com
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.bibleapi.com
    description: Production server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []
paths:
...
  /api/{version}/verse/{book}/{chapter}/{verse}:
    get:
      summary: Get single verse
      description: Returns a specific Bible verse. This endpoint is TRACKED for RSS feeds and writes to user's configured TXT/XML files.
      tags: [Bible]
      security:
        - bearerAuth: []
      parameters:
        - name: version
          in: path
          required: true
          schema:
            type: string
            enum: [kjv, tam]
          description: Bible version code
          example: kjv
        - name: book
          in: path
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 5
          description: Book abbreviation (e.g., JHN, PSA, GEN)
          example: JHN
        - name: chapter
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
            maximum: 150
          description: Chapter number
          example: 3
        - name: verse
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
            maximum: 176
          description: Verse number
          example: 16
      responses:
        200:
          description: Verse found
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      version_code:
                        type: string
                      book_short_name:
                        type: string
                      book_name:
                        type: string
                      chapter_num:
                        type: integer
                      verse_num:
                        type: integer
                      verse:
                        type: string
                      reference:
                        type: string
        400:
          description: Validation error (invalid chapter/verse)
        401:
          description: Authentication required
        403:
          description: Access denied - Version not assigned
        404:
          description: Verse not found
  /api/{version}/range/{book}/{startChapter}/{startVerse}/{endChapter}/{endVerse}:
    get:
      summary: Get verse range
      description: Returns a range of Bible verses
      tags: [Bible]
      security:
        - bearerAuth: []
      parameters:
        - name: version
          in: path
          required: true
          schema:
            type: string
        - name: book
          in: path
          required: true
          schema:
            type: string
        - name: startChapter
          in: path
          required: true
          schema:
            type: integer
        - name: startVerse
          in: path
          required: true
          schema:
            type: integer
        - name: endChapter
          in: path
          required: true
          schema:
            type: integer
        - name: endVerse
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Range of verses
        400:
          description: Validation error
        401:
          description: Authentication required
        403:
          description: Access denied
  /api/{version}/chapter/{book}/{chapter}:
    get:
      summary: Get complete chapter
      description: Returns all verses in a specific chapter
      tags: [Bible]
      security:
        - bearerAuth: []
      parameters:
        - name: version
          in: path
          required: true
          schema:
            type: string
        - name: book
          in: path
          required: true
          schema:
            type: string
        - name: chapter
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Complete chapter
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      version:
                        type: string
                      book:
                        type: string
                      book_name:
                        type: string
                      chapter:
                        type: integer
                      total_verses:
                        type: integer
                      max_verses_in_chapter:
                        type: integer
                      verses:
                        type: array
        400:
          description: Validation error
        401:
          description: Authentication required
        403:
          description: Access denied
  /api/{version}/search/{keyword}:
    get:
      summary: Search Bible verses
      description: Search for verses containing specific keywords
      tags: [Bible]
      security:
        - bearerAuth: []
      parameters:
        - name: version
          in: path
          required: true
          schema:
            type: string
        - name: keyword
          in: path
          required: true
          schema:
            type: string
          description: Search keyword or phrase
        - name: testament
          in: query
          schema:
            type: string
            enum: [Old, New]
          description: Filter by testament
        - name: book
          in: query
          schema:
            type: string
          description: Filter by book abbreviation
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
          description: Maximum number of results
      responses:
        200:
          description: Search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      version:
                        type: string
                      keyword:
                        type: string
                      total_results:
                        type: integer
                      results:
                        type: array
        401:
          description: Authentication required
        403:
          description: Access denied
...
  /api/validate:
    post:
      summary: Validate Bible reference
      description: Validate if a Bible reference is valid
      tags: [Bible]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - book
                - chapter
                - verse
              properties:
                book:
                  type: string
                chapter:
                  type: integer
                verse:
                  type: integer
      responses:
        200:
          description: Validation result
          content:
            application/json:
              schema:
                type: object
                properties:
                  valid:
                    type: boolean
                  error:
                    type: string
                  bookName:
                    type: string
                  maxVerses:
                    type: integer

  # ==================== RSS FEEDS ====================
...